Build time container for dependencies
Since the dep ensure step is quite costly to execute, a new
intermediate container is added to generate dependencies from
Gopkg.toml and Gopkg.lock files.

This avoids also the need of a go environment just for the
purpose of executing dep ensure, reducing the dependencies for
CI slaves.

The container definition is added in both Dockerfile and
Dockerfile-node verbatim, so the Docker builder should consider
the result the same given the same contents in Gopkg.toml and
Gopkg.lock files, avoiding dependencies recomputation and
retrieval and reusing the deps container image from Docker cache.

Also, since vendor is obtained from deps container, only pkg/
and cmd/ directories are copied in builder container. This way
the Docker cache is better reused in case non-source changes are
performed in the directory.

Change-Id: I5adaa7a4e23618a7336ca86921a988a273ae184e
Signed-off-by: Miguel Herranz <miguel@midokura.com>