CNI: Remove MAC annotation on DEL
For some reasons, a terminating Pod sometimes linger for days
after its sandbox is shut down.  Probably it's a bug in kubelet.
But even without bugs, it can happen with finalizers.
IP/MAC info in the MidoNet bridge for such a Pod can lead troubles
if the IP address is reused by another Pod.
Clearing the MAC annotation here fixes the issue because it would
make the controller remove the IP/MAC info.

Ref: MNA-1320
Signed-off-by: YAMAMOTO Takashi <yamamoto@midokura.com>
Change-Id: I86516083b69508475c77251a34b1b145c6566a96