Run unit tests when building docker images
This will run the unit tests at build time. To easier integration
with CI tools, a JUnit file is generated using go2xunit and copied
from the intermediate `builder` container. Since it is needed a
successful build to be able to recover the JUnit report, failing
tests are ignored in `builder` container build. The error is raised
in a subsequent `test_fail_checker` auxiliar container that repeats
the go2xunit conversion in 'error on fail mode'. In case the build
is a cross compilation for another arch, the tests will be not
executed, but the build process will succeed as there will be no
tests results (neither pass not fail).

Change-Id: I9f4a2a0ad83a7c42ca3e9be94fa61d072d4959c1
Signed-off-by: Miguel Herranz <miguel@midokura.com>