Improve docker cache usage on build time
A new intermediate build container is created that is build only
in case new dependencies need to be downloaded. This way changes in
the code or files at root repo directory do not invalidate the cache
forcing redownload of dependencies, which takes time and bandwidth
resources, specially for CI tools.

Change-Id: I315a3a0c5624eeb4553569079917e3f69c86d433
Signed-off-by: Miguel Herranz <miguel@midokura.com>